Update comment about the high bits in ECRecover v value (#1602)
The ECRecover precompile was already returning an empty vector for invalid v values,
but the comment wasn't clear about handling high bits. This commit updates the
comment to be more explicit about checking for any set high bits in the v value
(bytes 32-62 must be 0) and that v (byte 63) can only be 27 or 28.

Fixes TODO comment regarding "InvalidHighV-bits-1" test case.